cache workspace components
